[CodeStyle] Del obsolete code of partition_aggregation_node
