+++ be/src/common/config.h
@@ -337,8 +337,7 @@ namespace config {
 
     // for partition
     CONF_Bool(enable_partitioned_hash_join, "false")
-    CONF_Bool(enable_partitioned_aggregation, "false")
-    CONF_Bool(enable_new_partitioned_aggregation, "true")
+    CONF_Bool(enable_partitioned_aggregation, "true")
 
     // for kudu
     // "The maximum size of the row batch queue, for Kudu scanners."
